Moved gfw/common/beam/pipeline/config.py to gfw/common/pipeline_config.py to avoid need to download beam dependencies to use.
