+++ src/gfw/common/beam/pipeline/__init__.py
@@ -34,7 +34,6 @@ Modules
 """
 
 from .base import Pipeline
-from .config import PipelineConfig
 from .dag import Dag, DagFactory, LinearDag, LinearDagFactory
 from .factory import PipelineFactory
 
@@ -45,6 +44,5 @@ __all__ = [
     "LinearDag",
     "LinearDagFactory",
     "Pipeline",
-    "PipelineConfig",
     "PipelineFactory",
 ]
